fix: Simplify File Storage API - MEED-7762 - Meeds-io/meeds#2568 (#1004)
Prior to this change, the File Service API was marking the files as
deleted while it continue to serve it as it wasn't deleted until a
weekly job is run to cleanup files deleted for more than a month ago.
This change will delete this strategy and will delete the file right
away without differing the deletion. In addition, each file, even when
the same binary as another existing file, will have its own unique
checksum generated in order to ensure that each file is managed
independently from others.
In addition, this will fix a Branding API Bug which will remove the
`Logo` and/or `Favicon` File when updating the Branding information
without a change.

This bug detection was possible when the file service API has been
improved to:
1/ Not serve Files marked as deleted
2/ Delete the File instantly.

+++ component/portal/src/main/java/org/exoplatform/portal/branding/BrandingServiceImpl.java
@@ -905,6 +905,7 @@ public class BrandingServiceImpl implements BrandingService, Startable {
                        fileId,
                        settingKey);
     if (fileId != null
+        && StringUtils.isNotBlank(uploadId)
         && !StringUtils.equals(BRANDING_RESET_ATTACHMENT_ID, uploadId)) {
       // Cleanup old fileId
       fileService.deleteFile(fileId);
